requirements.txt
graphviz==0.20.3
matplotlib==3.9.0
numpy==1.24.3
opencv_python_headless==4.7.0.72
pandas==2.2.2
Pillow==10.3.0
scipy==1.13.1
scikit-image==0.23.2
tqdm==4.65.0